+++ CHANGELOG.md
@@ -4,7 +4,8 @@
  `connectionState == ConnectionState.none`
 * Add `isWaiting` getter to the `ReactiveModel` as a shortcut to :
  `connectionState == ConnectionState.waiting`
-* Add `errorHandler(BuildContext, dynamic)` parameter to `setState` method.
+* Add `onError(BuildContext, dynamic)` parameter to `setState` method.
+* Add `joinSingletonToNewData` parameter to `setState` method.
 * Refactor codes to remove bugs and to use Flutter 1.12 version.
 
 ## 1.8.0 (2019-12-06)